Feat: #108 VerificationRepository 인터페이스 패턴 리팩토링
- VerificationRepository를 인터페이스로 분리
- VerificationRepositoryImpl 구현체 생성
- RepositoryModule에서 인터페이스 주입 설정
- LoginViewModel에서 인터페이스 의존성 사용
